Move insight series picker to the shared components directory
