+++ client/web/src/enterprise/insights/components/code-insights-page/CodeInsightsPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { Page } from '../../../../components/Page'
-import { useUiFeatures } from '../../hooks/use-ui-features'
+import { useUiFeatures } from '../../hooks'
 
 import { CodeInsightsLimitAccessBanner } from './limit-access-banner/CodeInsightsLimitAccessBanner'
