+++ client/web/src/enterprise/insights/components/creation-ui/creation-ui-layout/CreationUiLayout.tsx
