+++ client/web/src/enterprise/insights/components/creation-ui/form-series/FormSeries.module.scss
