+++ client/web/src/enterprise/insights/components/creation-ui/form-series/components/form-color-input/FormColorInput.module.scss
