+++ client/web/src/enterprise/insights/components/creation-ui/form-series/components/form-series-input/get-pattern-type-filter.ts
