+++ client/web/src/enterprise/insights/components/creation-ui/form-series/components/series-card/SeriesCard.module.scss
