+++ client/web/src/enterprise/insights/components/creation-ui/form-series/index.ts
@@ -1,0 +1,3 @@
+export { FormSeries } from './FormSeries'
+export { createDefaultEditSeries } from './use-editable-series'
+export type { EditableDataSeries } from './types'
