+++ client/web/src/enterprise/insights/components/creation-ui/form-series/types.ts
@@ -1,0 +1,7 @@
+import { SearchBasedInsightSeries } from '../../../core'
+
+export interface EditableDataSeries extends SearchBasedInsightSeries {
+    valid: boolean
+    edit: boolean
+    autofocus: boolean
+}
