+++ client/web/src/enterprise/insights/components/creation-ui/index.ts
@@ -5,3 +5,5 @@ export { getSanitizedRepositories } from './sanitizers/repositories'
 
 export { CodeInsightDashboardsVisibility } from './CodeInsightDashboardsVisibility'
 export { CodeInsightTimeStepPicker } from './code-insight-time-step-picker/CodeInsightTimeStepPicker'
+export { FormSeries, createDefaultEditSeries } from './form-series'
+export type { EditableDataSeries } from './form-series'
