+++ client/web/src/enterprise/insights/components/creation-ui/live-preview/constants.ts
