+++ client/web/src/enterprise/insights/components/form/repositories-field/utils/get-suggestions-search-term.ts
@@ -1,4 +1,4 @@
-import { getSanitizedRepositories } from '../../../creation-ui-kit'
+import { getSanitizedRepositories } from '../../../creation-ui'
 
 interface SuggestionsSearchTermInput {
     value: string
