+++ client/web/src/enterprise/insights/components/form/validators.ts
@@ -2,9 +2,9 @@ import { Validator } from './hooks/useField'
 import { ValidationResult } from './hooks/useForm'
 
 /**
- * Validator for required form field which returns error massage
+ * Validator for required form field which returns error message
  * as a sign of invalid state.
- * */
+ */
 export const createRequiredValidator = <Value>(errorMessage: string): Validator<Value> => (value, validity) => {
     if (validity?.valueMissing) {
         return errorMessage
@@ -20,6 +20,6 @@ export const createRequiredValidator = <Value>(errorMessage: string): Validator<
 
 /**
  * Composes a few validators together and show first error for form field.
- * */
+ */
 export const composeValidators = <Value>(...validators: Validator<Value>[]): Validator<Value> => (value, validity) =>
     validators.reduce<ValidationResult>((error, validator) => error || validator(value, validity), undefined)
