+++ client/web/src/enterprise/insights/components/index.ts
@@ -3,7 +3,7 @@ export { CodeInsightsIcon } from '../../../insights/Icons'
 export * from './insights-view-grid'
 export * from './views'
 export * from './trancated-text/TruncatedText'
-export * from './creation-ui-kit'
+export * from './creation-ui'
 export * from './form'
 export * from './limited-access-label/LimitedAccessLabel'
 export * from './code-insights-page/CodeInsightsPage'
