+++ client/web/src/enterprise/insights/components/views/chart/index.ts
@@ -1,5 +1,5 @@
-export { SeriesChart } from './series/SeriesChart'
-export { CategoricalChart } from './categorical/CategoricalChart'
+export { SeriesChart, SeriesBasedChartTypes } from './series/SeriesChart'
+export { CategoricalChart, CategoricalBasedChartTypes } from './categorical/CategoricalChart'
 
 export type { SeriesChartProps } from './series/SeriesChart'
 export type { CategoricalChartProps } from './categorical/CategoricalChart'
