--- client/web/src/enterprise/insights/components/views/types.ts
@@ -1,8 +1,0 @@
-export enum SeriesBasedChartTypes {
-    Line,
-}
-
-export enum CategoricalBasedChartTypes {
-    Pie,
-    Bar,
-}
